requirements.txt
asgiref==3.2.7
astroid==2.4.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
Django==3.0.5
django-rq==2.3.2
djangorestframework==3.11.0
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.1.2
pylint==2.5.0
pytz==2020.1
redis==3.5.1
requests==2.23.0
rq==1.4.0
six==1.14.0
sqlparse==0.3.1
toml==0.10.0
typed-ast==1.4.1
urllib3==1.25.9
wrapt==1.12.1
